Add /completions
Port the concept of request_wrap from python - implement as a Lazy<T>
which takes a lambda to calculate the value and stores the result in an
optional<T>. To set the value we can just emplace the optional (though
we don't do this yet).

There's one issue which is that we're sending back fields which are
null; sould probably fix that but it's busywork typing to fix the json
std::optional<> conversion.