Implement basic guff for LSP
the idea here is to work out how to have independent message loops
within one coro framework. The magic that took me literally days to work
out is asio::async_initiate, which fundamentally calls a lambda, passing
it a callback. We then squirrel that callback into a struct (using
function2, which has a move_only_function: unique_function) and call
that callback when we get the response in the other loop.

You will not believe how long it took me to work all that out.